Apply cookiecutter to newly split project telemetry-tempest-plugin
This tempest plugin is being split out of the main ceilometer project in
accordance with Queens goal "Split Tempest Plugins into Separate
Repos/Projects"[1]. This patch applies the standard boilerplate files
for OpenStack projects so that it can stand on its own.

It contains tempest tests for Aodh, Gnocchi, Ceilometer and Panko projects.

[1] https://governance.openstack.org/tc/goals/queens/split-tempest-plugins.html