Revert "Remove autoscaling until gate issues sorted out"
This reverts commit 34e514a4824a5e47a895222fdb42f0f7d7b0123d.

In general, we'd want to check autoscaling from end to end.

Until now this plugin could only scale up the stack by adding
an instance. With this change, scaling down of instance is
also tested using a scaledown policy and an alarm which
is triggered if instance CPU usage falls below the threshold.

Depends-On: https://review.opendev.org/865386
Depends-On: https://review.opendev.org/868105
Depends-On: https://review.opendev.org/c/openstack/telemetry-tempest-plugin/+/904746

Co-authored-by: Yadnesh Kulkarni <ykulkarn@redhat.com>

Change-Id: I7f19dcaf327b636ef8328c6634d295414b7c8e65