Set aodh_threshold to "6000000000" in tempest.conf
For autoscaling granularity is set to 60 and the heat
stack is expected to scale out after instance reaches
CPU utilization higher than 10%.

According to this configuration the cpu time must be
10**9*60*(10/100) which is 6000000000 ns.

The default threshold for aodh alarms is "10000000"
which comes from ceilometer's devstack settings.

This change overrides the aodh_threshold value in
tempest.conf to the correct cpu time as per
cpu util and granularity values.

Signed-off-by: Yadnesh Kulkarni <ykulkarn@redhat.com>
Change-Id: Ic90613b70556d0ce5bcf64c2305a3351928bf83e