Disable GLOBAL_VENV and disable "telemetry-dsvm-integration-train"
* Devstack recently introduced the functionality to use a single venv[1]
but some devstack plugins in telemetry project are not compatible with
this change.

This disables the option as a short term workaround.

* Tempest CI job "telemetry-dsvm-integration-train" for Train release
fails due to
```
Obtaining file:///opt/stack/nova
nova requires Python '>=3.8' but the running Python is 3.6.9
```
Train release of Nova and Heat projects have reached EOL which are
fetched while running this job. If any issues are identified in these
releases fixing them might not be possible.

[1] https://review.opendev.org/c/openstack/devstack/+/558930

Change-Id: I01ed2c96395baf294aa84e4f989ab82f9b387fc4