Separate prometheus related tests
We have an issue, that tempest doesn't distinguish between
branches. New features aren't available on old branches and
so we can't execute testing, which involves sg-core or prometheus
on old branches. Separating sg-core and prometheus related tests
allows us to use tempest regex to execute those tests only on
branches, which have the required features.

This also fills in service_available requirements, to make it
clear which tests are runable on prometheus based storage
and which are runable on gnocchi based storage.

Change-Id: Ic5bea08295d8c85303c676706415e093625ee16f