Fix sg-core test scenario.
The scenario assumed, there is an image in glance, which may
not always be the case. This patch creates a new image based
on path configured in CONF.scenario.img_file. To add this
functionality the TestTelemetryIntegrationPrometheus class
was moved to inherit from tempest.scenario.manager.ScenarioTest.
Requirements on glance and ceilometer were added too. Ceilometer
was required even before this change, but the requirement check
was missing.

The scenario assumed the ceilometer pipeline interval is less
than 2 minutes. This patch adds a config option for this.
By default 300s is used, which is the default interval when
installing devstack. It's then changed to 15s in .zuul.yaml,
which is the interval used in our tests.

Change-Id: Ifae664540fc5a749b2965822267d7c54b34a77a5