Parametrize prometheus rate duration
Depending on the ceilometer and prometheus configuration, it
may be required to change the duration used by the prometheus
rate function in autoscaling queries. The prefered value
for stable testing should be <ceilometer_polling_interval> +
<prometheus_scrape_interval>. With other values (especially with
values less than the ceilometer polling interval), it's possible,
that the rate() function will just return 0 for ever.

Change-Id: Ifbb6da5c5be9db6cba5494b6d28b24d28967ec1c