Cleanup leftover stacks.
When the autoscaling scenario fails, gabbi doesn't execute the
rest of the steps in the scenario. This means the stack
used for the autoscaling test doesn't get deleted. The tempest
tearDown isn't able to delete the stack, which means we get
additional errors from the tearDown.

This patch adds a code which runs before the tearDown. It checks
if there is a stack running and it deletes it if necessary.

Change-Id: I7f88a29a296b213851e4987557903192938f31e4