Parametrize autoscaling instance load length
We need the duration for which the instance for autoscaling
testing is under load parametrized. Currently with a static
duration of 2 minutes the scenario might fail in situations
with a high ceilometer polling interval. In these situations
the load might already be finished before ceilometer polls
for the CPU metric, which results in a seemingly low CPU
utilization, because of which the stack doesn't ever scale
up and the scenario fails.

With this patch, the load lasts 2x (the ceilometer polling
interval + prometheus scrape interval),
which guarantees that at least on polling period
will be with 100% load, which will result in a high CPU
utilization registered by ceilometer.

As a sideeffect this might slightly speedup the scenario
in situations with a low ceilometer polling interval.

Change-Id: Ib673e67800d2d16d552e734fea89772a0e937105