Use server_group in Prom autoscaling scenario
Thanks to the recent developments in ceilometer and sg-core,
we can use server_group for grouping instances from the same
stack for autoscaling purposes. This is how the instances are
grouped in gnocchi based autoscaling. It's much easier for the
users to configure and it should be the prefered option when using
autoscaling.

For backwards compatibility with current stable branches I
added a "autoscaling_instance_grouping" config option. The old
way ("prefix") of instance grouping is used by default and so
tempest tests will continue working on stable branches. By
setting the option to "metadata", the new way of instance
grouping will be used. I'll set this setting in .zuul.yaml
of all telemetry repositories on master branches in follow-up
patches.

Change-Id: I2770e9d47b914941f938f63d92ab7868fe09d7b9