Add job with Aetos
Let's add a job, which has Aetos in front of Prometheus, but keep
the original job to ensure telemetry works both with Aetos and
with pure Prometheus. Let's also keep it non-voting for now until
the work proceeds a bit more and until we ensure the job works
and is stable. I'll add the job in Aetos in a little bit.

Depends-On: https://review.opendev.org/c/openstack/python-observabilityclient/+/953568

Change-Id: I2f4fc46a5a1e9621e3cd6871d120f7b6a1a2fe94
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>