Create network for integration scenarios
Current behavior creates a stack for the autoscaling scenario,
where it uses the first available network for creating servers.
We've seen failures in our downstream CI, which seem to be caused
by the scenario chosing ironic provisioning network.

With this patch a new network and a subnet are created at the
start of the integration scenarios. The network and subnet
is then deleted at the end of course.

Change-Id: Ib049bf96a345f64181cdca6616b44ba97e26d07a