Check service available flag for availability of gnocchi
telemetry-tempest-plugin provides the two options to toggle tests about
gnocchi.
 (1) [service_available] gnocchi
 (2) [telemetry_services] metric_backends

When (1) is set to False while (2) contains gnocchi, all gnocchi API
tests are skipped but the aodh API tests which requires gnocchi as
metric backend are still executed.

Although this combination might be misconfiguration, we assume that
the aodh API tests may likely fail in this case, so skip these tests in
this configuration pattern.

Change-Id: I078b93304dcb543727d715a4d7d995bf3839a70a