Use dummy resource type in tests
The `ceph_account` resource type is one of the actual resource types
created by ceilometer, and it should not be deleted after the tests
run. Create a resource type specifically used by tests.

Also retry deletion of the test archive policy because it can't be
deleted until all of the associated measures expire.

Change-Id: I0d423fc89858787fa970c9a38b5a6e40fc985c05
Co-Authored-By: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>
Signed-off-by: Takashi Kajinami <kajinamit@oss.nttdata.com>